photoevent/services/worker/requirements.txt
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
minio==7.2.3
httpx==0.26.0
Pillow==10.2.0
python-dotenv==1.0.0
